requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
langchain>=0.3.0
langgraph>=0.2.0
openai>=1.40.0
chromadb>=0.5.5
pypdf>=4.3.0
python-docx>=1.1.2
tiktoken>=0.7.0
pytest>=8.3.0
httpx>=0.27.0
